+++ wp-content/plugins/advisories-headers/src/Headers.php
@@ -6,12 +6,71 @@ class Headers
 {
 	private int $ttl_short = 1800;  // 30 mins..
 	private int $ttl_long = 86400;  // 1 day.
+	public const NONCE_NAME = 'csp';
 
 	public function register(): void
 	{
 		add_filter('wp_headers', [$this, 'addCacheControl']);
 		add_filter('wp_headers', [$this, 'addStrictTransportPolicy']);
-		add_filter('wp_headers', [$this, 'addContentSecurityPolicy']);
+		if (!is_admin()) {
+			add_filter('wp_headers', [$this, 'addContentSecurityPolicy']);
+		}
+		add_filter('wp_script_attributes', [$this, 'addCSPScriptAttributes'], 99999);
+		add_filter('wp_inline_script_attributes', [$this, 'addCSPScriptAttributes'], 99999);
+		add_filter('style_loader_tag', [$this, 'addCSPStyleAttributes'], 99999, 1);
+		add_action('wp_enqueue_scripts', [$this, 'removeGlobalStyles'], 99999, 0);
+		add_filter('wp_img_tag_add_auto_sizes', [$this, 'removeAutoImageSizes'], 99999, 0);
+	}
+
+	/**
+	 * Generate a nonce to use in a content security policy header.
+	 */
+	private function getCSPNonce(): string
+	{
+		return wp_create_nonce(self::NONCE_NAME);
+	}
+
+	/**
+	 * Remove style tag from WordPress core that defines contain-intrinsic-size.
+	 */
+	public function removeAutoImageSizes(): bool
+	{
+		return false;
+	}
+
+	/**
+	 * Remove enqueued styles from WordPress core to which we cannot add a nonce.
+	 */
+	public function removeGlobalStyles(): void
+	{
+		wp_dequeue_style('global-styles');
+		wp_dequeue_style('classic-theme-styles');
+	}
+
+	/**
+	 * Add a nonce to all <script> tags that are enqueued in Wordpress.
+	 */
+	public function addCSPScriptAttributes(array $attributes): array
+	{
+		if (!isset($attributes['nonce'])) {
+			$attributes['nonce'] = esc_attr($this->getCSPNonce());
+		}
+		return $attributes;
+	}
+
+	/**
+	 * Add a nonce to all <style> tags that are enqueued in Wordpress.
+	 *
+	 * Note that this misses some styles that are written directly into the
+	 * output buffer.
+	 */
+	public function addCSPStyleAttributes(string $tag): string
+	{
+		if (str_contains($tag, '<style') && !str_contains($tag, 'nonce=')) {
+			$nonce = $this->getCSPNonce();
+			$tag = str_replace('<style', '<style nonce="' . esc_attr($nonce) . '"', $tag);
+		}
+		return $tag;
 	}
 
 	/**
@@ -66,10 +125,10 @@ class Headers
 	{
 		$policy = [
 			"default-src 'self';",
-			"script-src 'self' 'unsafe-inline' data: https://plausible.io https://wordpress.org;",
+			"script-src 'self' 'nonce-" . esc_attr($this->getCSPNonce()) . "' data: https://plausible.io https://wordpress.org;",
 			"connect-src 'self' data: https://plausible.io https://wordpress.org;",
-			"img-src 'self' data: https://plausible.io https://wordpress.org;",
-			"style-src 'self' 'unsafe-inline';",
+			"img-src 'self' data: https://plausible.io https://wordpress.org https://secure.gravatar.com;",
+			"style-src 'self' 'nonce-" . esc_attr($this->getCSPNonce()) . "';",
 			"font-src 'self' data: https://wordpress.org;",
 			"object-src 'none';",  // <object> and <embed>
 			"media-src 'none';",  // <video>, <audio> and <track>